requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
inflection==0.5.1
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
